requirements.txt
# FastAPI + Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# MongoDB
pymongo==4.6.1
motor==3.3.2

# AI / DeepSeek
openai==1.12.0
numpy==1.26.3
scikit-learn==1.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# PDF parsing (for resumes)
PyPDF2==3.0.1
python-docx==1.1.0